refactor(data-service): remove package dependency cycles
