+++ yak-ops-business/yak-ops-business-data-service/src/main/java/io/yak/ops/business/dataservice/controller/v1/DataServiceRuntimeController.java
@@ -3,9 +3,9 @@ package io.yak.ops.business.dataservice.controller.v1;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import io.yak.framework.common.Result;
+import io.yak.ops.business.dataservice.domain.DataServiceQueryResponse;
 import io.yak.ops.business.dataservice.domain.InvocationRecord;
 import io.yak.ops.business.dataservice.execution.DataServiceInvoker;
-import io.yak.ops.business.dataservice.execution.DataServiceQueryResponse;
 import io.yak.ops.business.dataservice.observability.DataServiceCallLogReader;
 import io.yak.ops.business.dataservice.runtime.DataServiceRuntimePolicyManager;
 import io.yak.ops.business.dataservice.runtime.DataServiceRuntimeSnapshot;
