+++ yak-ops-business/yak-ops-business-data-service/src/main/java/io/yak/ops/business/dataservice/domain/DataServiceQueryResponse.java
@@ -1,4 +1,4 @@
-package io.yak.ops.business.dataservice.execution;
+package io.yak.ops.business.dataservice.domain;
 
 import java.util.List;
 import java.util.Map;
