+++ yak-ops-business/yak-ops-business-data-service/src/main/java/io/yak/ops/business/dataservice/query/DataServiceParameterNameReader.java
@@ -1,0 +1,9 @@
+package io.yak.ops.business.dataservice.query;
+
+import java.util.List;
+
+/** Read-side port for deriving the named parameters exposed by one Data Service SQL template. */
+public interface DataServiceParameterNameReader {
+
+  List<String> parameterNames(String sql);
+}
