+++ yak-ops-business/yak-ops-business-data-service/src/main/java/io/yak/ops/business/dataservice/query/DataServiceViewFactory.java
@@ -3,7 +3,6 @@ package io.yak.ops.business.dataservice.query;
 import io.yak.ops.business.dataservice.domain.DataServiceDefinition;
 import io.yak.ops.business.dataservice.domain.DataServiceSettings;
 import io.yak.ops.business.dataservice.domain.SourceReference;
-import io.yak.ops.business.dataservice.execution.DataServiceSqlCompiler;
 import io.yak.ops.business.datasource.config.ConditionalOnDataSourceEnabled;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -14,7 +13,7 @@ import org.springframework.stereotype.Component;
 public class DataServiceViewFactory {
 
   private static final String RUNTIME_PREFIX = "/api/v1/data-service/runtime";
-  private final DataServiceSqlCompiler sqlCompiler;
+  private final DataServiceParameterNameReader parameterNameReader;
 
   public DataServiceView view(DataServiceDefinition definition) {
     if (definition == null) return null;
@@ -23,7 +22,7 @@ public class DataServiceViewFactory {
     return new DataServiceView(
         definition.id(), settings.name(), settings.path(), RUNTIME_PREFIX + settings.path(),
         definition.runtimeSnapshot().dataSourceId(), definition.runtimeSnapshot().sql(),
-        sqlCompiler.parameterNames(definition.runtimeSnapshot().sql()), settings.maxRows(),
+        parameterNameReader.parameterNames(definition.runtimeSnapshot().sql()), settings.maxRows(),
         settings.timeoutSeconds(), settings.enabled(), definition.authMode().name(), settings.description(),
         source.sourceType(), source.sourceRef(), source.sourceRevisionId(), source.sourceRevisionNo(),
         definition.createTime(), definition.updateTime(), settings.paginationEnabled());
